# stl-verify/ts/ui/src/generated/openapi-schema.json
{
  "components": {
    "schemas": {
...
      "AllocationRiskCapitalResponse": {
        "description": "Per-allocation risk capital from the default model.",
        "properties": {
          "applied": {
            "description": "Whether the default model priced this allocation.",
            "title": "Applied",
            "type": "boolean"
          },
          "crr_pct": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Comparable capital-risk ratio (0-100). `null` when the allocation is unpriced.",
            "title": "Crr Pct"
          },
          "exposure_usd": {
            "description": "On-chain USD exposure of the allocation.",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Exposure Usd",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Model that produced the figure, or `null`.",
            "title": "Model"
          },
          "protocol_name": {
            "description": "Protocol the allocation sits in.",
            "title": "Protocol Name",
            "type": "string"
          },
          "receipt_token_id": {
            "description": "Surrogate id of the receipt token.",
            "title": "Receipt Token Id",
            "type": "integer"
          },
          "required_risk_capital_usd": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-allocation RRC (USD). `null` when the allocation is unpriced.",
            "title": "Required Risk Capital Usd"
          },
          "symbol": {
            "description": "Receipt-token symbol.",
            "title": "Symbol",
            "type": "string"
          },
          "unpriced_reason": {
            "anyOf": [
              {
                "const": "no_model",
                "type": "string"
              },
              {
                "enum": [
                  "share_data_missing",
                  "share_data_stale"
                ],
                "type": "string"
              },
              {
                "const": "price_data_missing",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Why the allocation is unpriced (`null` when `applied`): `no_model` (no default model applies), or `share_data_missing` / `share_data_stale` (a model applies but its pool-share lookup could not be resolved, e.g. a warm-up window or an un-indexed receipt token).",
            "title": "Unpriced Reason"
          }
        },
        "required": [
          "receipt_token_id",
          "symbol",
          "protocol_name",
          "exposure_usd",
          "applied"
        ],
        "title": "AllocationRiskCapitalResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PrimeRiskCapitalResponse": {
        "description": "Self-computed, model-derived capital metrics for a prime.\n\nIndependent of the upstream Star feed. `required_risk_capital_usd` is the\nsum of per-allocation RRC from the default model (`model`); it is **partial**\n(only allocations the model can price contribute) and **will not** match\nSky's dashboard. `modeled_pct` reports the priced share of exposure.",
        "properties": {
          "encumbrance_ratio": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Required / Total Risk Capital. `null` when total is absent or zero.",
            "title": "Encumbrance Ratio"
          },
          "exposure_usd": {
            "description": "\u03a3 priced receipt-token allocation exposure (USD).",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Exposure Usd",
            "type": "string"
          },
          "model": {
            "description": "Default RRC model used (e.g. `gap_sweep`).",
            "examples": [
              "gap_sweep"
            ],
            "title": "Model",
            "type": "string"
          },
          "modeled_exposure_usd": {
            "description": "Exposure the default model could price (USD).",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Modeled Exposure Usd",
            "type": "string"
          },
          "modeled_pct": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "`modeled_exposure_usd / exposure_usd` (0-1). `null` when exposure is zero.",
            "title": "Modeled Pct"
          },
          "per_allocation": {
            "description": "Per-allocation breakdown, newest-exposure first.",
            "items": {
              "$ref": "#/components/schemas/AllocationRiskCapitalResponse"
            },
            "title": "Per Allocation",
            "type": "array"
          },
          "prime_id": {
            "description": "Prime's 0x-prefixed ALM proxy address.",
            "title": "Prime Id",
            "type": "string"
          },
          "required_risk_capital_usd": {
            "description": "\u03a3 per-allocation RRC from the default model (USD).",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Required Risk Capital Usd",
            "type": "string"
          },
          "total_risk_capital_usd": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "On-chain SubProxy treasury balance (USD). `null` when absent.",
            "title": "Total Risk Capital Usd"
          }
        },
        "required": [
          "prime_id",
          "model",
          "exposure_usd",
          "required_risk_capital_usd",
          "modeled_exposure_usd",
          "per_allocation"
        ],
        "title": "PrimeRiskCapitalResponse",
        "type": "object"
      },
...
      "TimeSeriesResolution": {
        "description": "Allowed ISO-8601 durations for time-series downsampling.",
        "enum": [
          "PT1M",
          "PT5M",
          "PT15M",
          "PT1H",
          "PT6H",
          "P1D"
        ],
        "title": "TimeSeriesResolution",
        "type": "string"
      },
      "TimeSeriesWindow": {
        "description": "The resolved window and resolution actually applied to a request.\n\nEchoing this back lets consumers distinguish an empty result caused by the\nwindow from one caused by the absence of data.",
        "properties": {
          "from_timestamp": {
            "description": "Inclusive lower bound applied (UTC).",
            "format": "date-time",
            "title": "From Timestamp",
            "type": "string"
          },
          "interval_ms": {
            "description": "Resolution width in milliseconds.",
            "title": "Interval Ms",
            "type": "integer"
          },
          "resolution": {
            "$ref": "#/components/schemas/TimeSeriesResolution",
            "description": "Resolution applied (relevant when aggregated)."
          },
          "to_timestamp": {
            "description": "Inclusive upper bound applied (UTC).",
            "format": "date-time",
            "title": "To Timestamp",
            "type": "string"
          }
        },
        "required": [
          "from_timestamp",
          "to_timestamp",
          "resolution",
          "interval_ms"
        ],
        "title": "TimeSeriesWindow",
        "type": "object"
      },
...
      "TotalCapitalBucketResponse": {
        "description": "Last observed treasury balance within a single time bucket (LOCF gap-filled).",
        "properties": {
          "bucket_start": {
            "description": "Inclusive start of the time bucket (UTC).",
            "format": "date-time",
            "title": "Bucket Start",
            "type": "string"
          },
          "total_capital_usd": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Last observed SubProxy treasury USDS balance carried forward into the bucket (USD; USDS is dollar-pegged), serialized as a JSON string. `null` for leading buckets before the first observation.",
            "examples": [
              "36359440.25"
            ],
            "title": "Total Capital Usd"
          }
        },
        "required": [
          "bucket_start"
        ],
        "title": "TotalCapitalBucketResponse",
        "type": "object"
      },
      "TotalCapitalEnvelope": {
        "description": "Per-prime total-capital time series, gap-filled into buckets.",
        "properties": {
          "data": {
            "description": "Last treasury balance per time bucket.",
            "items": {
              "$ref": "#/components/schemas/TotalCapitalBucketResponse"
            },
            "title": "Data",
            "type": "array"
          },
          "mode": {
            "const": "aggregated",
            "description": "Always `aggregated`: a gap-filled time series.",
            "title": "Mode",
            "type": "string"
          },
          "window": {
            "$ref": "#/components/schemas/TimeSeriesWindow",
            "description": "The window and resolution applied to this response."
          }
        },
        "required": [
          "mode",
          "window",
          "data"
        ],
        "title": "TotalCapitalEnvelope",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Verify-side HTTP API for the STL pipeline.\n\nEndpoints expose primes (capital allocators), their allocations and debt, decoded protocol events, token catalog and pricing, and risk-capital computations (RRC, bad debt, breakdown).",
    "title": "stl-verify",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/primes/{prime_id}/risk-capital": {
      "get": {
        "description": "Compute the prime's capital metrics from on-chain data and the default RRC model (`gap_sweep`), with no dependency on the upstream Star feed. Returns exposure (priced receipt-token allocations), Total Risk Capital (on-chain treasury), Required Risk Capital (sum of per-allocation model RRC), encumbrance, a `modeled_pct` coverage figure, and a per-allocation breakdown. The figures are model-derived and partial (only allocations the model can price contribute Required Risk Capital) and will not match Sky's dashboard. A backed allocation whose pool-share lookup can't be resolved (e.g. a warm-up window or an un-indexed receipt token) is reported as unpriced (`applied=false` with an `unpriced_reason`) rather than failing the whole response. Returns `404` if the prime is unknown.",
        "operationId": "get_prime_risk_capital_v1_primes__prime_id__risk_capital_get",
        "parameters": [
          {
            "description": "A prime's 0x-prefixed ALM **proxy** address on one chain \u2014 not a prime identifier. A prime allocates through one proxy per chain; list them via `GET /v1/primes` and group by `prime_vault_address`.",
            "in": "path",
            "name": "prime_id",
            "required": true,
            "schema": {
              "description": "A prime's 0x-prefixed ALM **proxy** address on one chain \u2014 not a prime identifier. A prime allocates through one proxy per chain; list them via `GET /v1/primes` and group by `prime_vault_address`.",
              "title": "Prime Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrimeRiskCapitalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Self-computed prime risk capital",
        "tags": [
          "primes",
          "capital",
          "primes",
          "capital"
        ]
      }
    },
    "/v1/primes/{prime_id}/total-capital": {
      "get": {
        "description": "Return the prime's total capital over time, gap-filled (LOCF) into buckets. Total capital is the treasury USDS held in the prime's SubProxy wallet (USDS is dollar-pegged, so the balance is the USD figure); it matches the upstream Star `total_capital`. Returns `404` if the prime is unknown. Defaults to the last 24h; pass a window and `resolution` for longer ranges.",
        "operationId": "list_prime_total_capital_v1_primes__prime_id__total_capital_get",
        "parameters": [
          {
            "description": "A prime's 0x-prefixed ALM **proxy** address on one chain \u2014 not a prime identifier. A prime allocates through one proxy per chain; list them via `GET /v1/primes` and group by `prime_vault_address`.",
            "in": "path",
            "name": "prime_id",
            "required": true,
            "schema": {
              "description": "A prime's 0x-prefixed ALM **proxy** address on one chain \u2014 not a prime identifier. A prime allocates through one proxy per chain; list them via `GET /v1/primes` and group by `prime_vault_address`.",
              "title": "Prime Id",
              "type": "string"
            }
          },
          {
            "description": "Max buckets returned (default 100, max 500).",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "description": "Max buckets returned (default 100, max 500).",
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Inclusive lower timestamp bound (ISO-8601). Defaults to 24h before `to_timestamp`.",
            "in": "query",
            "name": "from_timestamp",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive lower timestamp bound (ISO-8601). Defaults to 24h before `to_timestamp`.",
              "title": "From Timestamp"
            }
          },
          {
            "description": "Inclusive upper timestamp bound (ISO-8601). Defaults to the current UTC time.",
            "in": "query",
            "name": "to_timestamp",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive upper timestamp bound (ISO-8601). Defaults to the current UTC time.",
              "title": "To Timestamp"
            }
          },
          {
            "description": "ISO-8601 duration resolution (for example `PT5M`, `PT1H`). Used for time-bucketing when `aggregate=true`; defaults to the finest resolution allowed for the window.",
            "in": "query",
            "name": "resolution",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TimeSeriesResolution"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO-8601 duration resolution (for example `PT5M`, `PT1H`). Used for time-bucketing when `aggregate=true`; defaults to the finest resolution allowed for the window.",
              "title": "Resolution"
            }
          },
          {
            "description": "When true, return time-bucketed aggregates instead of raw rows.",
            "in": "query",
            "name": "aggregate",
            "required": false,
            "schema": {
              "default": false,
              "description": "When true, return time-bucketed aggregates instead of raw rows.",
              "title": "Aggregate",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TotalCapitalEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prime total-capital (treasury) time series",
        "tags": [
          "primes",
          "capital"
        ]
      }
    },
...
  },
...
}